requirements.txt
streamlit
joblib
numpy
pandas
scikit-learn==1.3.2
PyPDF2==3.0.1
python-docx==0.8.11
